# 1.X版本_OpenAPI.json
{
  "swagger": "2.0",
  "info": {
    "description": "用于《电梯安检维修管理平台》项目的Api自动生成的Swagger2文档，可以导入ApiFox",
    "version": "1.0",
    "title": "电梯安检维修管理平台 API",
    "termsOfService": "http://localhost:8080/"
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "paths": {
    "/accessory": {
      "get": {
        "tags": [
          "配件对象功能接口"
        ],
        "summary": "通过ID查询单条数据",
        "operationId": "queryByIdUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "配件ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "post": {
        "tags": [
          "配件对象功能接口"
        ],
        "summary": "新增数据",
        "operationId": "addUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "accessory",
            "description": "accessory",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Accessory对象",
              "originalRef": "Accessory对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "put": {
        "tags": [
          "配件对象功能接口"
        ],
        "summary": "更新数据",
        "operationId": "editUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "accessory",
            "description": "accessory",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Accessory对象",
              "originalRef": "Accessory对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "delete": {
        "tags": [
          "配件对象功能接口"
        ],
        "summary": "通过主键删除数据",
        "operationId": "deleteByIdUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "配件ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/accessory/pagin": {
      "post": {
        "tags": [
          "配件对象功能接口"
        ],
        "summary": "分页查询",
        "operationId": "paginQueryUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "size",
            "in": "query",
            "description": "元素",
            "required": true,
            "type": "ref"
          },
          {
            "in": "body",
            "name": "accessory",
            "description": "accessory",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Accessory对象",
              "originalRef": "Accessory对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/elevator": {
      "get": {
        "tags": [
          "电梯对象功能接口"
        ],
        "summary": "通过ID查询单条数据",
        "operationId": "queryByIdUsingGET_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "电梯ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "post": {
        "tags": [
          "电梯对象功能接口"
        ],
        "summary": "新增数据",
        "operationId": "addUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "elevator",
            "description": "elevator",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Elevator对象",
              "originalRef": "Elevator对象"
            }
          },
          {
            "name": "files",
            "in": "formData",
            "description": "files",
            "required": false,
            "type": "array",
            "items": {
              "type": "file"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "put": {
        "tags": [
          "电梯对象功能接口"
        ],
        "summary": "更新数据",
        "operationId": "editUsingPUT_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "elevator",
            "description": "elevator",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Elevator对象",
              "originalRef": "Elevator对象"
            }
          },
          {
            "name": "files",
            "in": "formData",
            "description": "files",
            "required": false,
            "type": "array",
            "items": {
              "type": "file"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "delete": {
        "tags": [
          "电梯对象功能接口"
        ],
        "summary": "通过主键删除数据",
        "operationId": "deleteByIdUsingDELETE_1",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "电梯ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/elevator/pagin": {
      "post": {
        "tags": [
          "电梯对象功能接口"
        ],
        "summary": "分页查询",
        "operationId": "paginQueryUsingPOST_1",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "size",
            "in": "query",
            "description": "元素",
            "required": true,
            "type": "ref"
          },
          {
            "in": "body",
            "name": "elevator",
            "description": "elevator",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Elevator对象",
              "originalRef": "Elevator对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/elevatorImage": {
      "delete": {
        "tags": [
          "电梯图片对象功能接口"
        ],
        "summary": "通过主键删除数据",
        "operationId": "deleteByIdUsingDELETE_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "电梯图片ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/inspection": {
      "get": {
        "tags": [
          "检查报告对象功能接口"
        ],
        "summary": "通过ID查询单条数据",
        "operationId": "queryByIdUsingGET_2",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "检查报告ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "post": {
        "tags": [
          "检查报告对象功能接口"
        ],
        "summary": "新增数据",
        "operationId": "addUsingPOST_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "files",
            "in": "formData",
            "description": "files",
            "required": false,
            "type": "array",
            "items": {
              "type": "file"
            },
            "collectionFormat": "multi"
          },
          {
            "in": "formData",
            "name": "inspection",
            "description": "inspection",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inspection对象",
              "originalRef": "Inspection对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "put": {
        "tags": [
          "检查报告对象功能接口"
        ],
        "summary": "更新数据",
        "operationId": "editUsingPUT_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "files",
            "in": "formData",
            "description": "files",
            "required": false,
            "type": "array",
            "items": {
              "type": "file"
            },
            "collectionFormat": "multi"
          },
          {
            "in": "formData",
            "name": "inspection",
            "description": "inspection",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inspection对象",
              "originalRef": "Inspection对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "delete": {
        "tags": [
          "检查报告对象功能接口"
        ],
        "summary": "通过主键删除数据",
        "operationId": "deleteByIdUsingDELETE_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "检查报告ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/inspection/pagin": {
      "post": {
        "tags": [
          "检查报告对象功能接口"
        ],
        "summary": "分页查询",
        "operationId": "paginQueryUsingPOST_2",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束日期",
            "required": false,
            "type": "string"
          },
          {
            "name": "size",
            "in": "query",
            "description": "元素",
            "required": true,
            "type": "ref"
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "开始日期",
            "required": false,
            "type": "string"
          },
          {
            "in": "body",
            "name": "inspection",
            "description": "inspection",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inspection对象",
              "originalRef": "Inspection对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/inspectionImage": {
      "delete": {
        "tags": [
          "检查报告图片对象功能接口"
        ],
        "summary": "通过主键删除数据",
        "operationId": "deleteByIdUsingDELETE_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "维修报告图片ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
    "/inventory": {
      "get": {
        "tags": [
          "库存管理对象功能接口"
        ],
        "summary": "通过ID查询单条数据",
        "operationId": "queryByIdUsingGET_3",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "配件ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "put": {
        "tags": [
          "库存管理对象功能接口"
        ],
        "summary": "更新数据",
        "operationId": "editUsingPUT_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "inventory",
            "description": "inventory",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inventory对象",
              "originalRef": "Inventory对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/inventory/pagin": {
      "post": {
        "tags": [
          "库存管理对象功能接口"
        ],
        "summary": "分页查询",
        "operationId": "paginQueryUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "size",
            "in": "query",
            "description": "元素",
            "required": true,
            "type": "ref"
          },
          {
            "in": "body",
            "name": "inventoryVo",
            "description": "inventoryVo",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Inventory对象",
              "originalRef": "Inventory对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/location": {
      "get": {
        "tags": [
          "场所对象功能接口"
        ],
        "summary": "通过ID查询单条数据",
        "operationId": "queryByIdUsingGET_4",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "场所ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "post": {
        "tags": [
          "场所对象功能接口"
        ],
        "summary": "新增数据",
        "operationId": "addUsingPOST_3",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "location",
            "description": "location",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Location对象",
              "originalRef": "Location对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "put": {
        "tags": [
          "场所对象功能接口"
        ],
        "summary": "更新数据",
        "operationId": "editUsingPUT_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "location",
            "description": "location",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Location对象",
              "originalRef": "Location对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      },
      "delete": {
        "tags": [
          "场所对象功能接口"
        ],
        "summary": "通过主键删除数据",
        "operationId": "deleteByIdUsingDELETE_5",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "场所ID",
            "required": true,
            "type": "ref"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
    "/location/pagin": {
      "post": {
        "tags": [
          "场所对象功能接口"
        ],
        "summary": "分页查询",
        "operationId": "paginQueryUsingPOST_4",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "required": true,
            "type": "ref"
          },
          {
            "name": "size",
            "in": "query",
            "description": "元素",
            "required": true,
            "type": "ref"
          },
          {
            "in": "body",
            "name": "location",
            "description": "location",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Location对象",
              "originalRef": "Location对象"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result",
              "originalRef": "Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-order": "2147483647"
      }
    },
...
  },
  "definitions": {
    "Accessory对象": {
      "type": "object",
      "properties": {
        "accessoryName": {
          "type": "string",
          "description": "配件名称"
        },
        "accessoryNumber": {
          "type": "string",
          "description": "配件编号"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "description": "创建时间"
        },
        "id": {
          "type": "integer",
          "format": "int32",
          "description": "配件ID"
        },
        "isDeleted": {
          "type": "string",
          "description": "是否删除"
        },
        "remarks": {
          "type": "string",
          "description": "备注"
        },
        "specification": {
          "type": "string",
          "description": "规格型号"
        },
        "type": {
          "type": "string",
          "description": "配件类型"
        },
        "unit": {
          "type": "string",
          "description": "单位"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "description": "更新时间"
        }
      },
      "title": "Accessory对象",
      "description": "配件"
    },
...
    "Result": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "object"
        },
        "flag": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "Result"
    },
...
  }
}